config: add application factory and blueprint
