+++ MANIFEST.in
@@ -23,6 +23,7 @@ include COPYING
 include *.rst
 include *.sh
 include pytest.ini
+prune docs/_build
 recursive-include docs *.py
 recursive-include docs *.png
 recursive-include docs *.rst
